cloud: make *cloud.Client unit testable
In order to actually unit test a *cloud.Client, we needed some way
to pass in a mock connection. To facilitate that, there is a new
constructor (NewClientOpts) which uses functional options to control
the client, including optionally passing a (mockable) grpc client
in instead of passing in an address and relying on the constructor
to connect.

We have added a quick, simple test around unexpected EOFs in the
StreamLogs RPC to show it working, but much more should be unit
testable now.